[ew_encointer_utils] add more tests for compute meetup indexes
